fix carpetas clean architecture
