+++ src/main/java/com/SistemaDeReservas/Hotel/common/dto/HabitacionDto.java
@@ -1,9 +1,10 @@
-package com.SistemaDeReservas.Hotel.dto;
+package com.SistemaDeReservas.Hotel.common.dto;
 
-import com.SistemaDeReservas.Hotel.enums.EstadoHabitacion;
-import com.SistemaDeReservas.Hotel.enums.TipoHabitacion;
+import com.SistemaDeReservas.Hotel.domain.enums.EstadoHabitacion;
+import com.SistemaDeReservas.Hotel.domain.enums.TipoHabitacion;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -14,6 +15,7 @@ import lombok.ToString;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class HabitacionDto {
     private String numero;
     private TipoHabitacion tipo;
