+++ src/main/java/com/SistemaDeReservas/Hotel/common/dto/PisoDto.java
@@ -1,6 +1,7 @@
-package com.SistemaDeReservas.Hotel.dto;
+package com.SistemaDeReservas.Hotel.common.dto;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -11,6 +12,7 @@ import lombok.ToString;
 @ToString
 @NoArgsConstructor
 @AllArgsConstructor
+@Builder
 public class PisoDto {
     private String nombre;
     private String descripcion;
