+++ src/main/java/com/SistemaDeReservas/Hotel/common/dto/ReservaDto.java
@@ -1,10 +1,11 @@
-package com.SistemaDeReservas.Hotel.dto;
+package com.SistemaDeReservas.Hotel.common.dto;
 
 import java.time.LocalDate;
 
-import com.SistemaDeReservas.Hotel.enums.EstadoReserva;
+import com.SistemaDeReservas.Hotel.domain.enums.EstadoReserva;
 
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,6 +16,7 @@ import lombok.Setter;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+@Builder
 public class ReservaDto {
     private LocalDate fechaInicio;
     private LocalDate fechaFin;
