+++ src/main/java/com/SistemaDeReservas/Hotel/domain/enums/EstadoHabitacion.java
@@ -1,4 +1,4 @@
-package com.SistemaDeReservas.Hotel.enums;
+package com.SistemaDeReservas.Hotel.domain.enums;
 
 public enum EstadoHabitacion {
     DISPONIBLE,
