+++ src/main/java/com/SistemaDeReservas/Hotel/infraestructure/persistence/IPisoRepository.java
@@ -1,7 +1,8 @@
-package com.SistemaDeReservas.Hotel.repositories;
+package com.SistemaDeReservas.Hotel.infraestructure.persistence;
 
-import com.SistemaDeReservas.Hotel.models.Piso;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.SistemaDeReservas.Hotel.domain.model.Piso;
 
 public interface IPisoRepository  extends JpaRepository<Piso, Long> {
