+++ src/main/java/com/SistemaDeReservas/Hotel/infraestructure/persistence/IReservaRepository.java
@@ -1,8 +1,8 @@
-package com.SistemaDeReservas.Hotel.repositories;
+package com.SistemaDeReservas.Hotel.infraestructure.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.SistemaDeReservas.Hotel.models.Reserva;
+import com.SistemaDeReservas.Hotel.domain.model.Reserva;
 
 public interface IReservaRepository extends JpaRepository<Reserva, Long>{
